topic page and tc page
